[AS7-5789] HttpService as optional capability
