+++ osgi/integration/configadmin/src/main/java/org/jboss/as/osgi/configadmin/DomainModelListener.java
