+++ osgi/integration/http/pom.xml
@@ -1,0 +1,95 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+  ~ JBoss, Home of Professional Open Source.
+  ~ Copyright 2010, Red Hat, Inc., and individual contributors
+  ~ as indicated by the @author tags. See the copyright.txt file in the
+  ~ distribution for a full listing of individual contributors.
+  ~
+  ~ This is free software; you can redistribute it and/or modify it
+  ~ under the terms of the GNU Lesser General Public License as
+  ~ published by the Free Software Foundation; either version 2.1 of
+  ~ the License, or (at your option) any later version.
+  ~
+  ~ This software is distributed in the hope that it will be useful,
+  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
+  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+  ~ Lesser General Public License for more details.
+  ~
+  ~ You should have received a copy of the GNU Lesser General Public
+  ~ License along with this software; if not, write to the Free
+  ~ Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+  ~ 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+  -->
+
+<project
+    xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>org.jboss.as</groupId>
+        <artifactId>jboss-as-osgi-integration</artifactId>
+        <version>7.2.0.Alpha1-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>jboss-as-osgi-http</artifactId>
+    <name>JBoss Application Server: OSGi Integration Http</name>
+    <packaging>bundle</packaging>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.jboss.as</groupId>
+            <artifactId>jboss-as-web</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.jboss.as</groupId>
+            <artifactId>jboss-as-osgi-service</artifactId>
+        </dependency>
+    </dependencies>
+
+    <!-- Build -->
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.felix</groupId>
+                <artifactId>maven-bundle-plugin</artifactId>
+                <extensions>true</extensions>
+                <configuration>
+                    <instructions>
+                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
+                        <Bundle-Activator>org.jboss.as.osgi.httpservice.HttpBundleActivator</Bundle-Activator>
+                        <!-- Export-Package>
+                        	org.osgi.service.http;uses:="javax.servlet.http,javax.servlet";version="1.2.1",
+                        </Export-Package -->
+                        <Import-Package>
+                            javax.servlet;version="[3.0,4.0)", 
+                            javax.servlet.http;version="[3.0,4.0)",
+                            javax.management, 
+                            javax.security.auth.x500, 
+                            org.osgi.framework;version="[1.5,2.0)",
+                        	org.osgi.service.http;version="[1.2.1,1.2.2)",
+                            org.jboss.logging,
+                            org.jboss.modules,
+                            org.jboss.modules.management,
+                            org.jboss.modules.ref,
+                            org.jboss.msc,
+                            org.jboss.msc.inject,
+                            org.jboss.msc.service,
+                            org.jboss.msc.service.management,
+                            org.jboss.msc.value,
+                            org.jboss.osgi.framework,
+                            org.jboss.osgi.resolver,
+                            !org.apache.catalina*, 
+                            !org.apache.tomcat*, 
+                            !org.jboss.as.controller*, 
+                            !org.jboss.as.osgi*, 
+                            !org.jboss.as.server*, 
+                            !org.jboss.as.web*, 
+							</Import-Package>
+                    </instructions>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+</project>
