+++ osgi/integration/http/src/main/java/org/jboss/as/osgi/httpservice/IOUtils.java
