+++ osgi/integration/jmx/src/main/java/org/jboss/as/osgi/jmx/JMXExtension.java
