+++ osgi/integration/jpa/src/main/java/org/jboss/as/osgi/jpa/JPAExtension.java
