+++ osgi/integration/jpa/src/main/java/org/jboss/as/osgi/jpa/PersistenceUnitProcessor.java
