+++ osgi/integration/naming/src/main/java/org/jboss/as/osgi/naming/BundleContextBindingProcessor.java
