+++ osgi/integration/naming/src/main/java/org/jboss/as/osgi/naming/NamingExtension.java
