+++ osgi/integration/web/src/main/java/org/jboss/as/osgi/web/WebExtension.java
